refactor(assistant): extract reusable <AssistantChat>; /my/assistant is a thin host
